Add TimelineItem component
